+++ sdk/confidentialledger/azure-confidentialledger/CHANGELOG.md
@@ -10,6 +10,8 @@ GA Data Plane Python SDK for Confidential Ledger.
 ### Breaking Changes
 - Removed all models. Methods now return JSON directly.
 - `sub_ledger_id` fields are now named `collection_id`.
+- `azure.confidentialledger.identity_service` has been renamed to `azure.confidentialledger.certificate`.
+- `ConfidentialLedgerIdentityServiceClient` is now `ConfidentialLedgerCertificateClient`.
 
 ### Other Changes
 - Python 2.7 is no longer supported. Please use Python version 3.6 or later.